ElectromagneticFields.jl – Modules

ElectromagneticFields.code — Function

Generate code for evaluating analytic equilibria.

ElectromagneticFields.connection — Method

Returns the l-th component of the Levi-Civita connection

ElectromagneticFields.contravariant_to_covariant — Method

Returns the i-th component of the one-form corresponding to the vector v

ElectromagneticFields.contravariant_to_physical — Method

Returns the i-th component of the physical coordinate representation of the one-form α

ElectromagneticFields.covariant_to_contravariant — Method

Returns the i-th component of the vector corresponding to the one-form α

ElectromagneticFields.covariant_to_physical — Method

Returns the i-th component of the physical coordinate representation of the one-form α

ElectromagneticFields.crossproduct — Method

Returns the m-th component of the cross-product between the vectors v and w

ElectromagneticFields.generate_equilibrium_functions — Method

Generate functions for evaluating analytic equilibria.

ElectromagneticFields.hodge²¹ — Method

Returns the m-th component of the one-form corresponding to the two-form β

ElectromagneticFields.load_equilibrium — Function

Evaluate functions for evaluating analytic equilibria.

ElectromagneticFields.magnitude — Method

Returns the length of the vector v

ElectromagneticFields.normalize! — Method

Normalises the vector v in the metric g

ElectromagneticFields.normalize — Method

Normalises the vector v in the metric g

ElectromagneticFields.Γ — Method

Returns the Christoffel symbol Γⱼₖˡ