ElectromagneticFields.jl – Modules

ElectromagneticFields.ABC — Type

ABC equilibrium in (x,y,z) coordinates with covariant components of the vector potential given by

\[A (x,y,z) = \big( a \, \sin(z) + c \, \cos(y) , \, b \, \sin(x) + a \, \cos(z) , \, c \, \sin(y) + b \, \cos(x) \big)^T\]

resulting in the magnetic field $B(x,y,z) = A(x,y,z)$.

Parameters: a, b, c

ElectromagneticFields.AxisymmetricTokamakCartesian — Type

Axisymmetric tokamak equilibrium in (x,y,z) coordinates with covariant components of the vector potential given by

\[A (x,y,z) = \frac{1}{2} \frac{B_0}{q_0} \, \bigg( \frac{q_0 R_0 x z - r^2 y}{R^2} , \, \frac{q_0 R_0 y z + r^2 x}{R^2} , \, - q_0 R_0 \, \ln \bigg( \frac{R}{R_0} \bigg) \bigg)^T ,\]

resulting in the magnetic field with covariant components

\[B (x,y,z) = \frac{B_0}{q_0} \, \bigg( - \frac{q_0 R_0 y + x z}{R^2} , \, \frac{q_0 R_0 x - y z}{R^2} , \, \frac{R - R_0}{R} \bigg)^T ,\]

where $R = \sqrt{ x^2 + y^2 }$ and $r = \sqrt{ (R - R_0)^2 + z^2 }$.

Parameters:

R₀: position of magnetic axis

B₀: B-field at magnetic axis

q₀: safety factor at magnetic axis

ElectromagneticFields.AxisymmetricTokamakCylindrical — Type

Axisymmetric tokamak equilibrium in (R,Z,ϕ) coordinates with covariant components of the vector potential given by

\[A (R, Z, \phi) = \frac{B_0}{2} \, \bigg( R_0 \, \frac{Z}{R} , \, - R_0 \, \ln \bigg( \frac{R}{R_0} \bigg) , \, - \frac{r^2}{q_0} \bigg)^T ,\]

resulting in the magnetic field with covariant components

\[B (R, Z, \phi) = \frac{B_0}{q_0} \, \bigg( - \frac{Z}{R} , \, \frac{R - R_0}{R} , \, - q_0 R_0 \bigg)^T ,\]

where $r = \sqrt{ (R - R_0)^2 + Z^2 }$.

Parameters:

R₀: position of magnetic axis

B₀: B-field at magnetic axis

q₀: safety factor at magnetic axis

ElectromagneticFields.AxisymmetricTokamakToroidal — Type

Axisymmetric tokamak equilibrium in (r,θ,ϕ) coordinates with covariant components of the vector potential given by

\[A (r, \theta, \phi) = B_0 \, \bigg( 0 , \, \frac{r R_0}{\cos (\theta)} - \bigg( \frac{R_0}{\cos (\theta)} \bigg)^2 \, \ln \bigg( \frac{R}{R_0} \bigg) , \, - \frac{r^2}{2 q_0} \bigg)^T ,\]

resulting in the magnetic field with covariant components

\[B (r, \theta, \phi) = \frac{B_0}{q_0} \, \bigg( 0 , \, \frac{r^2}{R}, \, q_0 R_0 \bigg)^T ,\]

where $R = R_0 + r \cos \theta$.

Parameters:

R₀: position of magnetic axis

B₀: B-field at magnetic axis

q₀: safety factor at magnetic axis

ElectromagneticFields.EzCosZ — Type

Simple perturbation in electric field in (x,y,z) coordinates:

\[E(x,y,z) = E_0 \, \begin{pmatrix} 0 \\ 0 \\ \cos (2\pi z) \\ \end{pmatrix}\]

Parameters: E₀

ElectromagneticFields.Solovev — Type

Axisymmetric Solov'ev equilibra in (R/R₀,Z/R₀,phi) coordinates. Based on Cerfon & Freidberg, Physics of Plasmas 17, 032502, 2010.

Parameters:

R₀: position of magnetic axis

B₀: B-field at magnetic axis

ϵ: inverse aspect ratio

κ: elongation

δ: triangularity

a: free constant, determined to match a given beta value

ElectromagneticFields.SolovevQuadratic — Type

Quadratic Solov'ev equilibrium in (R,Z,phi) coordinates. Based on McCarthy, Physics of Plasmas 6, 3554, 1999.

Parameters:

R₀: position of magnetic axis

B₀: B-field at magnetic axis

a, b`: free constants

ElectromagneticFields.SolovevXpoint — Type

Axisymmetric Solov'ev equilibra with X-point in (R/R₀,Z/R₀,phi) coordinates. Based on Cerfon & Freidberg, Physics of Plasmas 17, 032502, 2010.

Parameters:

R₀: position of magnetic axis

B₀: B-field at magnetic axis

ϵ: inverse aspect ratio

κ: elongation

δ: triangularity

a: free constant, determined to match a given beta value

xₛₑₚ: x position of the X point

yₛₑₚ: y position of the X point

ElectromagneticFields.SymmetricQuadratic — Type

Symmetric quadratic equilibrium in (x,y,z) coordinates with covariant components of the vector potential given by

\[A (x,y,z) = \frac{B_0}{4} \big( - y \, (2 + x^2 + y^2) , \, x \, (2 + x^2 + y^2) , \, 0 \big)^T\]

resulting in the magnetic field with covariant components

\[B(x,y,z) = B_0 \, \begin{pmatrix} 0 \\ 0 \\ 1 + x^2 + y^2 \\ \end{pmatrix}\]

Parameters: B₀

ElectromagneticFields.ThetaPinch — Type

θ-pinch equilibrium in (x,y,z) coordinates with covariant components of the vector potential given by

\[A (x,y,z) = \frac{B_0}{2} \big( - y , \, x , \, 0 \big)^T\]

resulting in the magnetic field with covariant components

\[B (x,y,z) = \big( 0 , \, 0 , \, B_0 \big)^T .\]

Parameters: B₀: B-field at magnetic axis

ElectromagneticFields.generate_equilibrium_code — Function

Generate code for evaluating analytic equilibria.

ElectromagneticFields.generate_equilibrium_functions — Method

Generate functions for evaluating analytic equilibria.

ElectromagneticFields.get_vector_component — Method

Returns the i-th component of the vector corresponding to the one-form α

ElectromagneticFields.hodge²¹ — Method

Returns the m-th component of the one-form corresponding to the two-form β

ElectromagneticFields.load_equilibrium — Function

Evaluate functions for evaluating analytic equilibria.