Modules

ElectromagneticFields.jl – Modules

ElectromagneticFields.ABC — Type.

ABC equilibrium in (x,y,z) coordinates.

Parameters: a: b: c:

source

ElectromagneticFields.AxisymmetricTokamakCartesian — Type.

Axisymmetric tokamak equilibrium in (x,y,z) coordinates.

Parameters: R₀: position of magnetic axis B₀: B-field at magnetic axis q₀: safety factor at magnetic axis

source

ElectromagneticFields.AxisymmetricTokamakCylindrical — Type.

Axisymmetric tokamak equilibrium in (R,Z,ϕ) coordinates.

Parameters: R₀: position of magnetic axis B₀: B-field at magnetic axis q₀: safety factor at magnetic axis

source

ElectromagneticFields.AxisymmetricTokamakToroidal — Type.

Axisymmetric tokamak equilibrium in (r,θ,ϕ) coordinates.

Parameters: R₀: position of magnetic axis B₀: B-field at magnetic axis q₀: safety factor at magnetic axis

source

ElectromagneticFields.Solovev — Type.

Axisymmetric Solov'ev equilibra in (R/R₀,Z/R₀,phi) coordinates. Based on Cerfon & Freidberg, Physics of Plasmas 17, 032502, 2010.

Parameters: R₀: position of magnetic axis B₀: B-field at magnetic axis ϵ: inverse aspect ratio κ: elongation δ: triangularity a: free constant, determined to match a given beta value

source

ElectromagneticFields.SolovevQuadratic — Type.

Quadratic Solov'ev equilibrium in (R,Z,phi) coordinates. Based on McCarthy, Physics of Plasmas 6, 3554, 1999.

Parameters: R₀: position of magnetic axis B₀: B-field at magnetic axis a,b: free constants

source

ElectromagneticFields.SolovevXpoint — Type.

Axisymmetric Solov'ev equilibra with X-point in (R/R₀,Z/R₀,phi) coordinates. Based on Cerfon & Freidberg, Physics of Plasmas 17, 032502, 2010.

Parameters: R₀: position of magnetic axis B₀: B-field at magnetic axis ϵ: inverse aspect ratio κ: elongation δ: triangularity a: free constant, determined to match a given beta value xₛₑₚ: x position of the X point yₛₑₚ: y position of the X point

source

ElectromagneticFields.SymmetricQuadratic — Type.

Symmetric quadratic equilibrium in (x,y,z) coordinates.

\[B(x,y,z) = B_0 \, \begin{pmatrix} 0 \\ 0 \\ 1 + x^2 + y^2 \\ \end{pmatrix}\]

with electric field

\[E(x,y,z) = E_0 \, \begin{pmatrix} 0 \\ 0 \\ \cos (2\pi z) \\ \end{pmatrix}\]

Parameters: B₀, E₀

source

ElectromagneticFields.ThetaPinch — Type.

θ-pinch equilibrium in (x,y,z) coordinates.

Parameters: B₀: B-field at magnetic axis

source

ElectromagneticFields.load_equilibrium — Function.

Evaluate functions for evaluating analytic equilibria.

source

ElectromagneticFields.generate_equilibrium_code — Function.

Generate code for evaluating analytic equilibria.

source

ElectromagneticFields.generate_equilibrium_functions — Method.

Generate functions for evaluating analytic equilibria.

source

ElectromagneticFields.get_vector_component — Method.

Returns the i-th component of the vector corresponding to the one-form α

source

ElectromagneticFields.hodge²¹ — Method.

Returns the m-th component of the one-form corresponding to the two-form β

source